requirements_test.txt
# This file is originally from homeassistant/core and modified by pytest-homeassistant-custom-component.
# linters such as pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version

# types-* that have versions roughly corresponding to the packages they
# contain hints for available should be kept in sync with them

# -c homeassistant/package_constraints.txt
# -r requirements_test_pre_commit.txt
coverage==7.3.0
freezegun==1.2.2
mock-open==1.4.0
pydantic==1.10.12
pylint-per-file-ignores==1.2.1
pipdeptree==2.11.0
pytest-asyncio==0.21.0
pytest-aiohttp==1.0.4
pytest-cov==4.1.0
pytest-freezer==0.4.8
pytest-socket==0.6.0
pytest-test-groups==1.0.3
pytest-sugar==0.9.7
pytest-timeout==2.1.0
pytest-unordered==0.5.2
pytest-picked==0.4.6
pytest-xdist==3.3.1
pytest==7.3.1
requests_mock==1.11.0
respx==0.20.2
syrupy==4.2.1
tqdm==4.66.1
SQLAlchemy==2.0.15

paho-mqtt==1.6.1

numpy==1.23.2

# homeassistant==2023.7.3
pytest_homeassistant_custom_component